+++ com.beatwaves.responsible/Editor/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+// Since all editor stuff can be internal (and shouldn't be publicly used), use this hack.
+[assembly:InternalsVisibleTo("Responsible.EditorTests")]